api/requirements.txt
fastapi
uvicorn
pika
python-multipart
sqlalchemy
psycopg2-binary
pyjwt
passlib[bcrypt]
boto3
